requirements.txt
qrcode[pil]
pillow
reportlab
libsql-experimental
python-dotenv
